+++ tests/common.rs
@@ -1,10 +1,21 @@
-use staking_miner::prelude::Chain;
+use assert_cmd::cargo::cargo_bin;
+use codec::Decode;
+use sp_storage::StorageChangeSet;
+use staking_miner::{
+	opt::Chain,
+	prelude::{runtime, sp_core::Bytes, Hash, SubxtClient},
+};
 use std::{
 	io::{BufRead, BufReader, Read},
 	ops::{Deref, DerefMut},
 	process::{self, Child, ChildStderr, ChildStdout},
+	time::{Duration, Instant},
 };
+use subxt::rpc::rpc_params;
+use tokio::time::timeout;
 use tracing_subscriber::EnvFilter;
+
+const MAX_DURATION_FOR_SUBMIT_SOLUTION: Duration = Duration::from_secs(60 * 15);
 
 pub fn init_logger() {
 	let _ = tracing_subscriber::fmt()
@@ -46,7 +57,7 @@ pub fn run_staking_miner_playground() -> (KillChildOnDrop, String) {
 		process::Command::new("staking-miner-playground")
 			.stdout(process::Stdio::piped())
 			.stderr(process::Stdio::piped())
-			.args(&["--dev"])
+			.args(&["--dev", "--offchain-worker=Never"])
 			.env("RUST_LOG", "runtime=debug")
 			.spawn()
 			.unwrap(),
@@ -128,4 +139,76 @@ pub fn spawn_cli_output_threads(
 			}
 		}
 	});
+}
+
+pub enum Target {
+	PolkadotNode(Chain),
+	StakingMinerPlayground,
+}
+
+pub async fn test_submit_solution(target: Target) {
+	use runtime::runtime_types::pallet_election_provider_multi_phase::{
+		ElectionCompute, ReadySolution,
+	};
+
+	let (_drop, ws_url) = match target {
+		Target::PolkadotNode(chain) => run_polkadot_node(chain),
+		Target::StakingMinerPlayground => run_staking_miner_playground(),
+	};
+
+	let mut miner = KillChildOnDrop(
+		process::Command::new(cargo_bin(env!("CARGO_PKG_NAME")))
+			.stdout(process::Stdio::piped())
+			.stderr(process::Stdio::piped())
+			.args(&["--uri", &ws_url, "monitor", "--seed-or-path", "//Alice", "seq-phragmen"])
+			.spawn()
+			.unwrap(),
+	);
+
+	let (tx, mut rx) = tokio::sync::mpsc::unbounded_channel();
+	spawn_cli_output_threads(miner.stdout.take().unwrap(), miner.stderr.take().unwrap(), tx);
+
+	tokio::spawn(async move {
+		let r = rx.recv().await.unwrap();
+		log::info!("{}", r);
+	});
+
+	let api = SubxtClient::from_url(&ws_url).await.unwrap();
+	let now = Instant::now();
+
+	let key = Bytes(
+		runtime::storage()
+			.election_provider_multi_phase()
+			.queued_solution()
+			.to_root_bytes(),
+	);
+
+	let mut sub = api
+		.rpc()
+		.subscribe("state_subscribeStorage", rpc_params![vec![key]], "state_unsubscribeStorage")
+		.await
+		.unwrap();
+
+	let mut success = false;
+
+	while now.elapsed() < MAX_DURATION_FOR_SUBMIT_SOLUTION {
+		let x: StorageChangeSet<Hash> =
+			match timeout(MAX_DURATION_FOR_SUBMIT_SOLUTION, sub.next()).await {
+				Err(e) => panic!("Timeout exceeded: {:?}", e),
+				Ok(Some(Ok(storage))) => storage,
+				Ok(None) => panic!("Subscription closed"),
+				Ok(Some(Err(e))) => panic!("Failed to decode StorageChangeSet {:?}", e),
+			};
+
+		if let Some(data) = x.changes[0].clone().1 {
+			let solution: ReadySolution = Decode::decode(&mut data.0.as_slice())
+				.expect("Failed to decode storage as QueuedSolution");
+			eprintln!("solution: {:?}", solution);
+			assert!(solution.compute == ElectionCompute::Signed);
+			success = true;
+			break
+		}
+	}
+
+	assert!(success);
 }
